query e2e

+++ packages/cloud_firestore/cloud_firestore_platform_interface/lib/src/method_channel/method_channel_aggregate_query.dart
@@ -26,7 +26,7 @@ class MethodChannelAggregateQuery extends AggregateQueryPlatform {
         .aggregateQueryCount(_pigeonApp, _path, _pigeonParameters, source);
 
     return AggregateQuerySnapshotPlatform(
-      count: data,
+      count: data.toInt(),
     );
   }
 }
